+++ Sources/NostrSDK/Events/WalletConnect/WalletConnectNotificationEvent.swift
