+++ Sources/NostrSDK/Events/WalletConnect/WalletConnectRequestEvent.swift
