+++ Sources/NostrSDK/Events/WalletConnect/WalletConnectResponseEvent.swift
